agent-backend/requirements.txt
google-cloud-aiplatform[agent_engines,adk]>=1.38.0 # Use a specific version or >= minimum known good version
python-dotenv>=0.19.0